Remove comments and shows the right names for the tests.
Now became clear that only when the number of samples is 2048
or more the "better_solution" advantages are aparent.
The proportion of improvement can be seen as roughly logarithmic when
samples go to 4096 and (~16%) and above (~8% for 8192 and ~4% for 16384).
For a number of samples below 2048 the "trivial_solution" performs better
than the "better_solution".